move active cat to redux
