+++ src/redux/actions/activeCategoryActions.js
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+
+const changeActiveCategory = createAction('activeCategory/change');
+export default { changeActiveCategory };
