+++ src/redux/reducers/activeCategoryReducer.js
@@ -1,0 +1,7 @@
+import { createReducer } from '@reduxjs/toolkit';
+import actions from '../actions/activeCategoryActions';
+const activeCategory = createReducer('', {
+    [actions.changeActiveCategory]:(_, { payload }) => payload,
+})
+
+export default activeCategory;
